add support for half hour tide offsets

+++ db/migrate/20181005133059_create_condition_tides.rb
@@ -1,8 +1,8 @@
 class CreateConditionTides < ActiveRecord::Migration[5.2]
   def change
     create_table :condition_tides do |t|
-      t.integer :position_low_high, default: [], array: true
-      t.integer :position_high_low, default: [], array: true
+      t.float :position_low_high, default: [], array: true
+      t.float :position_high_low, default: [], array: true
       t.string :size, default: [], array: true
       t.timestamps
     end
