+++ db/schema.rb
@@ -34,8 +34,8 @@ ActiveRecord::Schema.define(version: 2018_10_05_225605) do
   end
 
   create_table "condition_tides", force: :cascade do |t|
-    t.integer "position_low_high", default: [], array: true
-    t.integer "position_high_low", default: [], array: true
+    t.float "position_low_high", default: [], array: true
+    t.float "position_high_low", default: [], array: true
     t.string "size", default: [], array: true
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
